Load the map lazily
The map at the bottom of the screen is very important, but is way below
the fold. This commit makes Chrome-based browsers load the iframe
lazily. This means that the first view of the page will be fast, but as
the user scrolls down, the map will load before it is visible.